ReportLab 3.1.44 has had quite a few minor tweaks. A heap of tests were affected in different ways.
Form rendering was changed and default values in text fields weren't appearing until focused. To fix the following test we have to modify the AcroForm from Document.process.

    tag-textField

Underlined Japanese characters now visually align with characters instead of the full character width. The test below was affected and the expected output PDF had to be updated.

    rml-examples-031-japanese

QR codes have been "improved" but the ability to set the bar stroke width and colour have been removed. The test below was affected and the expected output PDF had to be updated.

    tag-barCodeFlowable

When used in some locations the <font> tag was loading fonts incorrectly and the many tests below were loading Courier-Bold instead of Courier. All of them had to have their expected output PDFs updated.

    tag-alias
    tag-blockTable-1
    tag-blockTable-2
    tag-blockTable-3
    tag-blockTable-4
    tag-blockTable-5
    tag-blockTable-6
    tag-blockTable-7
    tag-blockTable-8
    tag-blockTable-9
    tag-blockTable-10
    tag-blockTable-bulkData
    tag-blockTableStyle
    tag-blockTableStyle-2
    tag-bookmark
    tag-fixedSize
    tag-illustration
    tag-imageAndFlowables
    tag-imageAndFlowables-svg
    tag-img
    tag-indent
    tag-place
    tag-plugInFlowable
    tag-pre
    tag-pto
    tag-spacer
    tag-xpre